feat(editor): prove the pure geometry and budget policies with Kani
Add Kani harnesses over the editor-memory budget and ledger accounting, the
minimap fit functions, the adaptive-shell layout and breakpoint, the
workspace-sidebar width presets, and the side-by-side preview-width clamp
(N2 of the formal-verification programme), in two new shards,
core-memory-policy and core-geometry-policies.

Two defects the harnesses found are fixed failing-first:
- a non-finite stored sidebar-width fraction (NaN, -inf) resolved to Large;
  it now resolves to the default, Comfy;
- the native-slider fit returned an infinite height for a source map near
  f64::MAX tall.

To make the policies checkable, budget selection takes the least-recently-used
page per step instead of sorting (same prefix), the ledger's accounting moves
into a private ResidencyTotals value, and the preview-width clamp moves into
ui/markdown_preview/policy.rs as whole-pixel integer arithmetic, with its
below-3-sp floor stated in the spec. Pure geometry and budget policy modules
now deny clippy::float_arithmetic, admitting fractional values only through
reasoned function-level expectations, and check-workflow-boundaries enforces
the attribute.

+++ crates/lushtext-core/src/ui/window/geometry/execution.rs
@@ -41,13 +41,14 @@ use libadwaita::prelude::AdwApplicationWindowExt;
 use crate::config::keys;
 use crate::ui::sidebar::width_preset::WorkspaceSidebarWidthPreset;
 
-use super::super::imp::{PREVIEW_LAYOUT_EDITOR, PREVIEW_MIN_WIDTH_SP};
+use super::super::imp::PREVIEW_LAYOUT_EDITOR;
 use super::policy::{
     self, AdaptiveShellInputs, OPEN_BUTTON_BREAKPOINT_MAX_WIDTH_SP,
     PROPERTIES_SIDEBAR_MIN_WIDTH_SP, PropertiesPresentation, WORKSPACE_SIDEBAR_MIN_WIDTH_SP,
     derive_adaptive_shell_layout, desired_properties_fraction, properties_breakpoint_condition,
     workspace_breakpoint_condition,
 };
+use crate::ui::markdown_preview::PREVIEW_MIN_WIDTH_SP;
 
 pub(in crate::ui::window) fn configure_split_views(
     workspace_split_view: &libadwaita::OverlaySplitView,
@@ -77,8 +78,8 @@ pub(in crate::ui::window) fn configure_split_views(
     preview_layout_view.set_layout_name(PREVIEW_LAYOUT_EDITOR);
     preview_split_view.set_sidebar_position(gtk4::PackType::End);
     preview_split_view.set_sidebar_width_unit(libadwaita::LengthUnit::Sp);
-    preview_split_view.set_min_sidebar_width(PREVIEW_MIN_WIDTH_SP);
-    preview_split_view.set_max_sidebar_width(PREVIEW_MIN_WIDTH_SP);
+    preview_split_view.set_min_sidebar_width(f64::from(PREVIEW_MIN_WIDTH_SP));
+    preview_split_view.set_max_sidebar_width(f64::from(PREVIEW_MIN_WIDTH_SP));
     preview_split_view.set_pin_sidebar(true);
     preview_split_view.set_enable_show_gesture(false);
     preview_split_view.set_enable_hide_gesture(false);
